use adaptor_test.go
